Make Moonraker utility module
